requirements_prod.txt
asgiref==3.6.0
Django==3.2.16
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
django_environ==0.9.0
python-dotenv==0.21.1
django-debug-toolbar==3.8.1